+++ src/si/acceleration.rs
@@ -58,8 +58,8 @@ quantity! {
 
 #[cfg(test)]
 macro_rules! test {
-    ($V:ty, $P:path) => {
-        use $P::*;
+    ($V:ident) => {
+        use ::si::$V::*;
         use ::si::acceleration as a;
         use ::si::length as l;
         use ::si::time as t;
@@ -108,11 +108,11 @@ macro_rules! test {
 mod tests {
     #[cfg(feature = "f32")]
     mod f32 {
-        test!(f32, ::si::f32);
+        test!(f32);
     }
 
     #[cfg(feature = "f64")]
     mod f64 {
-        test!(f64, ::si::f64);
+        test!(f64);
     }
 }