+++ CHANGELOG.md
@@ -10,6 +10,12 @@
 ### Security
 -->
 
+## [Unreleased]
+
+### Added
+ * [#11](https://github.com/iliekturtles/uom/issues/11) Floating point classification methods
+   `classify`, `is_finite`, `is_infinte`, `is_nan`,  and `is_normal` implemented for `Quantity`.
+
 ## [v0.14.0] — 2017-05-30
 
 ### Added